🐛 fix: fix price type integer to long

+++ src/main/java/umc/th/juinjang/api/note/shared/service/SharedNoteCommandService.java
@@ -128,7 +128,7 @@ public class SharedNoteCommandService {
 	@Transactional
 	public void createSharedNote(Member member, Long noteId, SharedNotePostRequest request) {
 		Integer rewardPencilCount = 0;
-		Integer price = 0;
+		Long price = 0L;
 
 		Optional<SharedNote> latestSharedNote = sharedNoteFinder.findLatestByLimjangId(noteId);
 		if (latestSharedNote.isPresent()) {
@@ -161,12 +161,12 @@ public class SharedNoteCommandService {
 				}
 			}
 			rewardPencilCount = 7;
-			price = 10;
+			price = 10L;
 		}
 		//사진 공유 안함 체크 or 임장노트에 사진이 없으면
 		else if (request.isImageShared() == Boolean.TRUE || limjang.getImageList().isEmpty()) {
 			rewardPencilCount = 2;
-			price = 5;
+			price = 5L;
 		}
 		limjang.updateRewardPencil(rewardPencilCount);
 		noteUpdater.save(limjang);
